==> 2_summary.Rmd <==
---
title: "Long-term mortality of judges: summary statistics and plots"
date: "`r format(Sys.time(), '%d %B %Y')`"
output: word_document
---

```{r setup, include=FALSE}
# using formatting in Word document (see above)
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, error=FALSE, comment='', dpi=400)
options(width=1000) # Wide pages
options(scipen=999) # avoid scientific presentation
library(dplyr)
library(tidyr)
library(janitor)
library(flextable)
library(Epi) # to make Lexis work
library(ggplot2)
g.theme = theme_bw() + theme(panel.grid.minor = element_blank())

# from 0_read_judges.R
load('data/analysis_ready.RData')
year_fraction = as.numeric(meta$censor.date - as.Date('2022-01-01')) / 365.25
```

# Lexis diagram

```{r, fig.width=7}
# from 1_lexis_diagram.R
load('results/lexis.RData')
par(mai=c(0.8, 0.8, 0.05, 0.05), las=1)
plot(judgesL, type='l', ylim=c(25, 106), ylab='Age, years', xlab='Year')
points(filter(judgesL, status=='Dead'), type='p', ylim=c(25, 106), pch=1, col='red', cex=0.5) # just dead
#
```

The Lexis diagram uses all the available data on judges (`r paste(format(range(judgesL$at_risk_date), '%Y'), collapse=' to ', sep='')`), so including data before the life tables were available in 1850.

# Number of new judges per year

```{r, fig.width=7}
# calculate number per year and age
to_plot = mutate(judges, 
                 year = as.numeric(format(at_risk_date, '%Y')), # year of entry
                 agein = as.numeric(difftime(at_risk_date, birth_date, units='days')/365.25)) %>%
  group_by(year) %>%
  summarise(n = n(), 
            mean = mean(agein), 
            sd = sd(agein),
            sem = sd / sqrt(n),
            z = qnorm(0.975),
            lower = mean - z*sem, # CI for the mean
            upper = mean + z*sem) %>%
  mutate(denom = ifelse(year == 2022, year_fraction, 1), # fraction of year
         n = n / denom) # adjust counts
#
label = data.frame(year=1850 + 1, n=120, text='Life table\ndata available') # plus one to move slightly to right
#
gplot = ggplot(data=to_plot, aes(x=year, y=n))+
  geom_line(col='dodgerblue', size=1.05)+
  g.theme+
  geom_vline(xintercept=1850, lty=2)+
  geom_text(data = label, aes(x=year, y=n, label=text), adj = 0, col='grey22')+
  xlab("Year")+
  ylab('Number of new judges')
gplot
```

The final year (2022) is adjusted for only being a fraction of a full year.

# Age at nomination over time

```{r, fig.width=7}
label = data.frame(year=1850 + 1, mean=68, lower=40, upper=40, text='Life table\ndata available') # plus one to move slightly to right
gplot = ggplot(data=to_plot, aes(x=year, y=mean, ymin=lower, ymax=upper))+
  geom_line(col='indianred1', size=1.05)+
  geom_ribbon(alpha=0.2)+
  geom_vline(xintercept=1850, lty=2)+
  geom_text(data = label, aes(x=year, y=mean, label=text), adj = 0, col='grey22')+
  g.theme+
  xlab("Year")+
  ylab('Age (years)')
gplot
```

The plot shows the mean age at nomination (red line) and a 95% confidence interval for the mean (shaded area). The confindence interval does not exist in some years as there were no judges appointed.


# Death counts per year

```{r, fig.width=7}
# calculate death number per year and age at death
to_plot_death = filter(judges, !is.na(death_date)) %>% # must have died
          mutate(year = as.numeric(format(death_date, '%Y')), # year of entry
                 ageout = as.numeric(difftime(death_date, birth_date, units='days')/365.25)) %>% # age at death
  group_by(year) %>%
  summarise(n = n(), 
            mean = mean(ageout), 
            sd = sd(ageout),
            sem = sd / sqrt(n),
            z = qnorm(0.975),
            lower = mean - z*sem, # CI for the mean
            upper = mean + z*sem) %>%
  mutate(denom = ifelse(year == 2022, year_fraction, 1), # fraction of year
         n = n / denom) # adjust counts
#
label = data.frame(year=1850 + 1, n=40, text='Life table\ndata available') # plus one to move slightly to right
#
gplot = ggplot(data=to_plot_death, aes(x=year, y=n))+
  geom_line(col='darkorange', size=1.05)+
  g.theme+
  geom_vline(xintercept=1850, lty=2)+
  geom_text(data = label, aes(x=year, y=n, label=text), adj = 0, col='grey22')+
  xlab("Year")+
  ylab('Number of deaths')
gplot
```

The final year (2022) is adjusted for only being a fraction of a full year.

# Age at death over time

```{r, fig.width=7}
label = data.frame(year=1850 + 1, mean=10, lower=40, upper=40, text='Life table\ndata available') # plus one to move slightly to right
gplot = ggplot(data=to_plot_death, aes(x=year, y=mean, ymin=lower, ymax=upper))+
  geom_line(col='darkorchid', size=1.05)+
  geom_ribbon(alpha=0.2)+
  geom_vline(xintercept=1850, lty=2)+
  geom_text(data = label, aes(x=year, y=mean, label=text), adj = 0, col='grey22')+
  coord_cartesian(ylim=c(0, NA))+
  g.theme+
  xlab("Year")+
  ylab('Age (years)')
gplot
```

The plot shows the mean age at death (red line) and a 95% confidence interval for the mean (shaded area). The confindence interval does not exist in some years as there were no judges appointed.


# Summary statistics

For the summary statistics below, we only use judges who were at risk from 1850.

```{r}
judges = filter(judges,
                at_risk_date >= as.Date('1850-01-01')) %>%
  mutate(follow = as.numeric((death_censor_date - at_risk_date)/365.25), # follow-up in years
         year = as.numeric(format(at_risk_date, '%Y')), # year of entry
         agein = as.numeric(difftime(at_risk_date, birth_date, units='days')/365.25)) # age at risk
stats = select(judges, nid, follow, year, agein) %>%
  pivot_longer(cols = -nid, names_to = 'variable') %>%
  group_by(variable) %>%
  summarise(min = round(min(value)), 
            q1 = round(quantile(value, 0.25)),
            q2 = round(quantile(value, 0.5)),
            q3 = round(quantile(value, 0.75)), 
            max = round(max(value))) %>%
  mutate(variable = case_when(
    variable == 'agein' ~ 'Age at nomination (years)',
    variable == 'follow' ~ 'Follow-up in years',
    variable == 'year' ~ 'Year of nomination',
  ) )
# table
ftab = flextable(stats) %>%
  theme_box() %>%
  autofit() %>%
  colformat_double(j=2:6, big.mark='', digits=0)
ftab
```

The total years of follow up is `r format(round(sum(judges$follow)),big.mark=',')`. There were `r sum(is.na(judges$death_date)==FALSE)` deaths.

```{r}
gender = judges %>%
  tabyl(gender) %>%
  mutate(percent = round(percent*100))
ftab = flextable(gender) %>%
  theme_box() %>%
  autofit()
ftab
```
  

# Average age over time

Here we examine the average age of the cohort over time.

```{r, fig.width=7}
# Calculate SMR and average age (code stolen from 1_SMR.R)
judges = rename(judges,
                'Status' = 'status', # rename to suit politicians code
         'Sex' = 'gender')
#
SMR = ages = NULL
follow.cut = 100 # longest follow-up
  # Calculate follow-up time depending on years of follow-up
  to.SMR = mutate(judges, 
                  follow.time = as.numeric(difftime(death_censor_date, at_risk_date, units='days')/365.25), # observed follow-up time in years from entry to death (or censoring)
                  death_datex = ifelse(follow.time > follow.cut, at_risk_date + (follow.cut*365.25), death_censor_date), # censor time at cut-off time if observed time is longer than time to death
                  death_datex = as.Date(death_datex, origin='1970-01-01'),
                  Statusx = ifelse(follow.time > follow.cut, 'Living', 'Dead'), # change status to living if still alive at this time
                  Statusx = ifelse(death_datex >= meta$censor.date, Status, Statusx), # revert back to previous status if beyond censor date
                  death_datex = ifelse(death_datex >= meta$censor.date, death_censor_date, death_datex), # revert back to previous time if beyond censor date
                  death_datex = as.Date(death_datex, origin='1970-01-01')) %>%
    select(-death_date, -Status) %>% # over-write Status and date-of-death
    rename('death_date' = 'death_datex', 'Status' = 'Statusx')
  
  ## Inner loop of calculating SMR
  # merge a year of judges with a year of data
  for (year in meta$Syear:meta$Eyear){ 
    # get life table for this year
    this.life = filter(life.table, Year==year) %>% 
      dplyr::select(Age, Sex, qx, imputed)
    # get relevant judges
    year.start = as.Date(paste(year, '-01-01', sep="")) # start of year
    year.end = as.Date(paste(year, '-12-31', sep="")) # end of year
    # calculate expected deaths
    exp = filter(to.SMR, at_risk_date <= year.end & death_date >= year.start) %>% # entered some time during the year and were alive at the start of the year
      mutate(age = floor(as.numeric(difftime(year.start, birth_date, units='days')/365.25)),  # calculate age at start of year ('current' age)
             age = ifelse(age > meta$Oldest, meta$Oldest, age)) # change any ages over limit of life table
    # calculate average age of judges
    av.age = summarise(exp, mean=mean(age)) %>%
      mutate(follow=follow.cut, year = year)
    ages = rbind(ages, av.age)
    # fractional first year (could not get this to work in dplyr)
    index = exp$entry.year == year
    exp$frac = 1 
    if(any(index)){exp$frac[index] = as.numeric(difftime(year.end, exp$at_risk_date[index], units='days'))/365.25}
    # merge with life table by age and sex
    exp = left_join(exp, this.life, by=c('age'='Age', 'Sex'='Sex')) %>% 
      summarise(expected=sum(qx*frac)) # expected deaths, fractional time in first year
    
    # calculate observed deaths
    obs = filter(to.SMR, death_date>=year.start & death_date<=year.end & Status=='Dead') %>% # died this year (not censored)
      summarise(observed=n())
    # calculate SMR and concatenate; add imputed (yes/no) flag for life tables
    frame = data.frame(follow=follow.cut, year=year, O=obs$observed, E=exp$expected, imputed=this.life$imputed[1])
    SMR = rbind(SMR, frame)
  } # end of inner loop 
SMR = mutate(SMR, SMR=O/E) # calculate SMR
```

```{r, fig.width=7}
gplot = ggplot(data=ages, aes(x=year, y=mean))+
  geom_line(col='darkseagreen4', size=1.05)+
  g.theme+
  xlab("Year")+
  ylab('Age (years)')
gplot
```

<!--- There was possibly a policy change in the upper age limit judges ages around around 1975? Searching on the web I can find: "Wisconsin used to have a mandatory retirement age for judges and justices. From 1955 to 1978, judges and justices had to retire at age 70." --->

# Observed and expected deaths

```{r, fig.width=7}
to_plot = select(SMR, year, O, E) %>%
  pivot_longer(-year) %>%
  mutate(name = ifelse(name=='O', 'Observed', 'Expected'))
gplot = ggplot(data = to_plot, aes(x=year, y=value, col=name))+
  geom_line(size=1.05)+
  scale_color_manual(NULL, values=c('grey','darkorange'))+
  g.theme+
  xlab("Year")+
  ylab("Death counts")+
  theme(legend.position=c(0.12,0.85))
gplot
```

# Standardised mortality ratio

```{r, fig.width=7}
gplot = ggplot(data = SMR, aes(x=year, y=SMR))+
  geom_hline(yintercept=1, lty=2, col='dark red')+
  geom_line(size=1.05, col='deepskyblue3')+
  g.theme+
  xlab("Year")+
  coord_cartesian(ylim=c(0,4)) # avoid crazy numbers
gplot
```
